pyproject.toml
[project]
name = "langchain-campusx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "arxiv>=2.4.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.1",
    "duckduckgo-search>=8.1.1",
    "embeddings>=0.0.8",
    "faiss-cpu>=1.13.2",
    "frozendict>=2.4.7",
    "google-generativeai>=0.8.6",
    "graphviz>=0.21",
    "groq>=0.37.1",
    "huggingface-hub>=0.36.1",
    "langchain>=1.2.8",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.8",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "lxml>=6.0.2",
    "numpy>=2.4.2",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "protobuf>=5.29.6",
    "pydantic>=2.12.5",
    "pymed>=0.8.9",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
    "sentencepiece>=0.2.1",
    "streamlit>=1.54.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=4.57.6",
    "wikipedia>=1.4.0",
]
